[feat]

+++ app/src/main/java/com/jeju/nanaland/domain/request/festival/GetMonthlyFestivalListRequest.kt
@@ -1,8 +1,8 @@
 package com.jeju.nanaland.domain.request.festival
 
 data class GetMonthlyFestivalListRequest(
-    val page: Long,
-    val size: Long,
+    val page: Int,
+    val size: Int,
     val addressFilterList: List<String>,
     val startDate: String?,
     val endDate: String?
