+++ app/src/main/java/com/jeju/nanaland/domain/request/nature/GetNatureListRequest.kt
@@ -6,7 +6,7 @@ data class GetNatureListRequest(
     @SerializedName("addressFilterList")
     val addressFilterList: List<String>,
     @SerializedName("page")
-    val page: Long,
+    val page: Int,
     @SerializedName("size")
-    val size: Long
+    val size: Int
 )
