+++ app/src/main/java/com/jeju/nanaland/ui/component/listscreen/filter/parts/FilteredItemCount.kt
@@ -11,7 +11,7 @@ import com.jeju.nanaland.util.ui.ComponentPreview
 
 @Composable
 fun FilteredItemCount(
-    count: Long
+    count: Int
 ) {
     Text(
         text = "$count " + getString(R.string.common_건),
